refactor: Move the Symfony classes to a Bridge namespace.
The current layout makes sense from the library consumer perspective,
but often I need to touch the actual bridge between the Symfony native
API and the library new API. In order to ease this task, I find helpful
to move those classes under a common `Bridge` namespace.

Note that it's mostly private (or should be private) classes anyway so I
do not expect to impact the consumer.